Start running tests on AppVeyor

+++ test/utils.py
@@ -3,6 +3,7 @@ import os.path
 import signal
 import time
 import difflib
+import platform
 
 
 def fail(msg):
@@ -57,9 +58,16 @@ def run_for(delay, command):
     """
     print('Going to run this for {}s: "{}"'.format(delay, ' '.join(command)))
     with open(os.devnull, 'w') as FNULL:
-        process = subprocess.Popen(command, stdout=FNULL, stderr=FNULL)
-        time.sleep(delay)
-        process.send_signal(signal.SIGINT)
+        # Windows wants a different treatment, see this issue:
+        # https://stackoverflow.com/questions/7085604/
+        if platform.system() == 'Windows':
+            wrapper_command = "start python3 windows_test_wrapper.py " + str(delay) + " '" + json.dumps(command) + "'"
+            print("Running the following command: \"" + wrapper_command + "\"")
+            subprocess.Popen(wrapper_command, shell=True)
+        else:
+            process = subprocess.Popen(command, stdout=FNULL, stderr=FNULL)
+            time.sleep(delay)
+            process.send_signal(signal.SIGINT)
 
 
 def check_fixture(name):
